+++ src/system/schedule/mod.rs
@@ -188,13 +188,13 @@ impl Schedule<stage::Null> {
     }
 }
 
-impl<'a, S> Schedule<S> {
+impl<S> Schedule<S> {
     pub(crate) fn run<R, SFI, SVI, PFI, PVI, SP, SI, SQ, PP, PI, PQ>(
         &mut self,
-        world: &'a mut World<R>,
+        world: &mut World<R>,
     ) where
         R: Registry,
-        S: Stages<'a, R, SFI, SVI, PFI, PVI, SP, SI, SQ, PP, PI, PQ>,
+        S: Stages<R, SFI, SVI, PFI, PVI, SP, SI, SQ, PP, PI, PQ>,
     {
         self.stages.run(
             // SAFETY: The pointer provided here is unique, being created from a mutable reference.