+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ### Changed
 - `Send` and `Sync` implementations of `World` now only require the registry to implement `Registry + Send` and `Registry + Sync` respectively.
 - Lifetime in `System` and `ParSystem` traits has been moved to the `Views` associated type.
-- The generic lifetime on the `system::schedule::RawTask` trait has been removed.
+- The generic lifetimes on the `system::schedule::RawTask` and `system::Stages` traits have been removed.
 
 ## 0.3.0 - 2022-10-28
 ### Added